+++ packages/next/src/data/server/fetchHookData.ts
@@ -2,6 +2,8 @@ import {
 	AbstractFetchStrategy,
 	EndpointParams,
 	FetchOptions,
+	FetchResponse,
+	FetchStrategyCacheConfig,
 	FilterDataOptions,
 	LOGTYPE,
 	PostParams,
@@ -9,10 +11,11 @@ import {
 } from '@headstartwp/core';
 import { GetServerSidePropsContext, GetStaticPropsContext } from 'next';
 import { serializeKey } from '@headstartwp/core/react';
-import deepmerge from 'deepmerge';
+import { all as merge } from 'deepmerge';
 import { PreviewData } from '../../handlers/types';
 import { convertToPath } from '../convertToPath';
 import { getSiteFromContext } from './getSiteFromContext';
+import defaultCacheHandler from './cache';
 
 /**
  * The supported options for {@link fetchHookData}
@@ -32,6 +35,11 @@ export interface FetchHookDataOptions<P = unknown, T = unknown> {
 	 * Optional. If set, will forward fetch options to the fetch strategy
 	 */
 	fetchStrategyOptions?: Partial<FetchOptions>;
+
+	/**
+	 * Controls server-side caching
+	 */
+	cache?: FetchStrategyCacheConfig;
 }
 
 function isPreviewRequest<P>(params: P, urlParams: P): params is P & PostParams {
@@ -47,6 +55,97 @@ function isPreviewRequest<P>(params: P, urlParams: P): params is P & PostParams 
 			: isPreviewRequestBasedOnNonUrlParams;
 
 	return isPreviewRequest;
+}
+
+/**
+ * Prepares all the things for fetchHookData
+ *
+ * @param fetchStrategy The fetch strategy to use. Typically this is exposed by the hook e.g: `usePosts.fetcher()`
+ * @param ctx The Next.js context, either the one from `getServerSideProps` or `getStaticProps`
+ * @param options See {@link FetchHookDataOptions}
+ *
+ * @returns The various things fetchHookData needs
+ */
+export function prepareFetchHookData<T = unknown, P extends EndpointParams = EndpointParams, R = T>(
+	fetchStrategy: AbstractFetchStrategy<T, P, R>,
+	ctx: GetServerSidePropsContext<any, PreviewData> | GetStaticPropsContext<any, PreviewData>,
+	options: FetchHookDataOptions<P, T> = {},
+) {
+	const { sourceUrl, integrations, cache: globalCacheConfig } = getSiteFromContext(ctx);
+
+	const defaultCacheTTL = 5 * 60;
+	const cacheConfig = merge<FetchStrategyCacheConfig>([
+		{
+			enabled: false,
+			ttl: defaultCacheTTL,
+			cacheHandler: defaultCacheHandler,
+		},
+		globalCacheConfig ?? {},
+		options.cache ?? {},
+	]);
+
+	const params: Partial<P> = options?.params || {};
+
+	fetchStrategy.setBaseURL(sourceUrl);
+
+	let path: string[] = [];
+
+	if (ctx.params) {
+		path = Array.isArray(ctx.params.path) ? ctx.params.path : [ctx.params.path || ''];
+	}
+
+	if (integrations?.polylang?.enable && (ctx.locale || ctx.defaultLocale)) {
+		params.lang = ctx.locale ?? ctx.defaultLocale;
+	}
+
+	const stringPath = convertToPath(path);
+	const defaultParams = fetchStrategy.getDefaultParams();
+	const urlParams = fetchStrategy.getParamsFromURL(stringPath, params);
+
+	const finalParams = merge([defaultParams, urlParams, params]) as Partial<P>;
+
+	const enabled = cacheConfig?.enabled ?? false;
+
+	const isCacheEnabled =
+		typeof enabled === 'function'
+			? enabled({
+					fetchStrategy,
+					params: finalParams,
+					fetchStrategyOptions: options.fetchStrategyOptions,
+					path: stringPath,
+					cacheHandler: cacheConfig.cacheHandler ?? defaultCacheHandler,
+				})
+			: enabled;
+
+	const shouldSkipCache = ctx.preview;
+
+	const shouldCache = isCacheEnabled && !shouldSkipCache;
+	const ttl = cacheConfig.ttl ?? defaultCacheTTL;
+
+	const cacheTTL =
+		typeof ttl === 'function'
+			? ttl({
+					fetchStrategy,
+					params: finalParams,
+					fetchStrategyOptions: options.fetchStrategyOptions,
+					path: stringPath,
+					cacheHandler: cacheConfig.cacheHandler ?? defaultCacheHandler,
+				})
+			: ttl;
+
+	return {
+		cacheKey: fetchStrategy.getCacheKey(finalParams),
+		params: finalParams,
+		urlParams,
+		path: stringPath,
+		cache: {
+			enabled: shouldCache,
+			ttl: cacheTTL,
+			cacheHandler: cacheConfig?.cacheHandler,
+			beforeSet: cacheConfig?.beforeSet,
+			afterGet: cacheConfig?.afterGet,
+		},
+	};
 }
 
 /**
@@ -67,13 +166,13 @@ function isPreviewRequest<P>(params: P, urlParams: P): params is P & PostParams 
  *	} catch (e) {
  *		return handleError(e, context);
  *	}
- * }
+ * }´
  * ```
  *
  * @param fetchStrategy The fetch strategy to use. Typically this is exposed by the hook e.g: `usePosts.fetcher()`
  * @param ctx The Next.js context, either the one from `getServerSideProps` or `getStaticProps`
  * @param options See {@link FetchHookDataOptions}
- *
+ 
  * @returns An object with a key of `data` and a value of the fetched data.
  *
  * @category Next.js Data Fetching Utilities
@@ -83,43 +182,29 @@ export async function fetchHookData<T = unknown, P extends EndpointParams = Endp
 	ctx: GetServerSidePropsContext<any, PreviewData> | GetStaticPropsContext<any, PreviewData>,
 	options: FetchHookDataOptions<P, T> = {},
 ) {
-	const { sourceUrl, integrations, debug, preview } = getSiteFromContext(ctx);
-	const params: Partial<P> = options?.params || {};
+	const {
+		cacheKey: key,
+		params,
+		urlParams,
+		path,
+		cache,
+	} = prepareFetchHookData(fetchStrategy, ctx, options);
 
-	fetchStrategy.setBaseURL(sourceUrl);
-
-	let path: string[] = [];
-
-	if (ctx.params) {
-		path = Array.isArray(ctx.params.path) ? ctx.params.path : [ctx.params.path || ''];
-	}
-
-	if (integrations?.polylang?.enable && (ctx.locale || ctx.defaultLocale)) {
-		params.lang = ctx.locale ?? ctx.defaultLocale;
-	}
-
-	const stringPath = convertToPath(path);
-	const defaultParams = fetchStrategy.getDefaultParams();
-	const urlParams = fetchStrategy.getParamsFromURL(stringPath, params);
-
-	const finalParams = deepmerge.all([defaultParams, urlParams, params]) as Partial<P>;
-
-	// we don't want to include the preview params in the key
-	const key = fetchStrategy.getCacheKey(finalParams);
+	const { debug, preview } = getSiteFromContext(ctx);
 
 	if (debug?.devMode) {
 		log(LOGTYPE.INFO, `[fetchHookData] key for  ${key.url}`, key);
 	}
 
 	if (
-		isPreviewRequest(finalParams, urlParams) &&
+		isPreviewRequest(params, urlParams) &&
 		typeof ctx.preview !== 'undefined' &&
 		typeof ctx.previewData !== 'undefined'
 	) {
-		finalParams.id = ctx.previewData.id;
-		finalParams.revision = ctx.previewData.revision;
-		finalParams.postType = ctx.previewData.postType;
-		finalParams.authToken = ctx.previewData.authToken;
+		params.id = ctx.previewData.id;
+		params.revision = ctx.previewData.revision;
+		params.postType = ctx.previewData.postType;
+		params.authToken = ctx.previewData.authToken;
 
 		if (debug?.requests) {
 			log(LOGTYPE.DEBUG, 'Preview request detected, using preview data', ctx.previewData);
@@ -138,16 +223,65 @@ export async function fetchHookData<T = unknown, P extends EndpointParams = Endp
 		}
 	}
 
-	const data = await fetchStrategy.fetcher(
-		fetchStrategy.buildEndpointURL(finalParams),
-		finalParams,
-		{
+	let data: FetchResponse<R> | null = null;
+	const cacheKey = serializeKey(key);
+
+	if (cache.enabled && cache.cacheHandler) {
+		data = (await cache.cacheHandler.get(cacheKey)) as FetchResponse<R>;
+
+		if (data) {
+			if (debug?.devMode) {
+				log(LOGTYPE.INFO, `[fetchHookData] cache hit for ${cacheKey}`);
+			}
+
+			if (typeof cache.afterGet === 'function') {
+				data = await cache.afterGet(
+					{
+						fetchStrategy,
+						params,
+						fetchStrategyOptions: options.fetchStrategyOptions,
+						path,
+						cacheHandler: cache.cacheHandler,
+					},
+					data,
+				);
+			}
+		} else if (debug?.devMode) {
+			log(LOGTYPE.INFO, `[fetchHookData] cache miss for ${cacheKey}`);
+		}
+	}
+
+	if (!data) {
+		data = await fetchStrategy.fetcher(fetchStrategy.buildEndpointURL(params), params, {
 			// burst cache to skip REST API cache when the request is being made under getStaticProps
 			// if .req is not available then this is a GetStaticPropsContext
 			burstCache: typeof (ctx as GetServerSidePropsContext).req === 'undefined',
 			...options.fetchStrategyOptions,
-		},
-	);
+		});
+
+		if (cache.enabled && cache.cacheHandler) {
+			if (debug?.devMode) {
+				log(LOGTYPE.INFO, `[fetchHookData] cache store for ${cacheKey}`);
+			}
+
+			if (typeof cache.beforeSet === 'function') {
+				data = await cache.beforeSet(
+					{
+						fetchStrategy,
+						params,
+						fetchStrategyOptions: options.fetchStrategyOptions,
+						path,
+						cacheHandler: cache.cacheHandler,
+					},
+					data,
+				);
+			}
+
+			await cache.cacheHandler.set(cacheKey, data, cache.ttl);
+		}
+	} else {
+		data.isCached = true;
+	}
 
 	if (debug?.devMode) {
 		log(LOGTYPE.INFO, `[fetchHookData] data.pageInfo for ${key.url}`, data.pageInfo);
@@ -155,7 +289,7 @@ export async function fetchHookData<T = unknown, P extends EndpointParams = Endp
 
 	const normalizedData = fetchStrategy.normalizeForCache(
 		fetchStrategy.filterData(data, options.filterData as unknown as FilterDataOptions<R>),
-		finalParams,
+		params,
 	);
 
 	let additionalCacheObjects;
@@ -164,14 +298,14 @@ export async function fetchHookData<T = unknown, P extends EndpointParams = Endp
 		additionalCacheObjects = normalizedData.additionalCacheObjects.map((cacheObject) => ({
 			...cacheObject,
 			key: serializeKey(cacheObject.key),
-			isMainQuery: fetchStrategy.isMainQuery(stringPath, params),
+			isMainQuery: fetchStrategy.isMainQuery(path, params),
 		}));
 	}
 
 	return {
 		...normalizedData,
-		key: serializeKey(key),
-		isMainQuery: fetchStrategy.isMainQuery(stringPath, params),
+		key: cacheKey,
+		isMainQuery: fetchStrategy.isMainQuery(path, params),
 		additionalCacheObjects: additionalCacheObjects || null,
 	};
 }