+++ packages/next/CHANGELOG.md
@@ -1,5 +1,75 @@
 # @headstartwp/next
 
+## 1.4.0
+
+### Minor Changes
+
+- cce64d2e: Introducing fetch strategy caching.
+
+### Patch Changes
+
+- 8d3a0f83: Normalize TTL to secs
+- 669441f3: Fix: cache handler types, add isCached property
+- 498fb394: caching: pass `fetchStrategyOptions` and `path` to cache functions.
+- bdcc37f0: Make cache.enabled optional
+- Updated dependencies [8d3a0f83]
+- Updated dependencies [669441f3]
+- Updated dependencies [cce64d2e]
+- Updated dependencies [498fb394]
+- Updated dependencies [bdcc37f0]
+  - @headstartwp/core@1.4.0
+
+## 1.4.0-next.4
+
+### Patch Changes
+
+- 8d3a0f83: Normalize TTL to secs
+- Updated dependencies [8d3a0f83]
+  - @headstartwp/core@1.4.0-next.4
+
+## 1.4.0-next.3
+
+### Patch Changes
+
+- bdcc37f0: Make cache.enabled optional
+- Updated dependencies [bdcc37f0]
+  - @headstartwp/core@1.4.0-next.3
+
+## 1.4.0-next.2
+
+### Patch Changes
+
+- 669441f3: Fix: cache handler types, add isCached property
+- Updated dependencies [669441f3]
+  - @headstartwp/core@1.4.0-next.2
+
+## 1.4.0-next.1
+
+### Patch Changes
+
+- 498fb394: caching: pass `fetchStrategyOptions` and `path` to cache functions.
+- Updated dependencies [498fb394]
+  - @headstartwp/core@1.4.0-next.1
+
+## 1.4.0-next.0
+
+### Minor Changes
+
+- cce64d2e: Introducing fetch strategy caching.
+
+### Patch Changes
+
+- Updated dependencies [cce64d2e]
+  - @headstartwp/core@1.4.0-next.0
+
+## 1.3.4
+
+### Patch Changes
+
+- 2c7313ac: Introducing `prepareFetchHookData` and making it public as it can be useful for projects customizing `fetchHookData` behavior.
+- Updated dependencies [5d5a856a]
+  - @headstartwp/core@1.3.4
+
 ## 1.3.3
 
 ### Patch Changes